add the function videoMultiThread()(unstable)
process the video with multi-thread.The main thread is just show the
video,and the thread TL_FRAME detects the TLs. I copy the video frame
and give the copy to TL_FRAME thread to process in order to avoid the
interference between the two threads. But this version is unstable and
will be crashed for a while